data set sucks - sorry
the position of the quotes does not make sence to me. Either wrap field
in quotes and use quotes within, or don't. Wrapping multiple fields
within quotes is not systematic. Moreover, adding single double quotes
to the price in itself does not make sence.
The data would have to be prepared before uploading or handled in the
api. Since I cannot find any systematics, preparing the data within
either front or backend seems like a task leading nowhere. The system
could be broken with line 15.